# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gestión de Proyectos",
        "version": "1.0.0",
        "title": "Gazelook",
        "contact": {
            "email": "benitez.leonardo06@gmail.com"
        }
    },
    "host": "localhost:4000",
    "basePath": "/api",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
        "/user": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Crea el usuario junto con la direccion y cuenta de usuario",
                "description": "Creación de Usuario, Direccion y Cuenta de Usuario",
                "operationId": "createUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateUser"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Retorna la lista de usuarios",
                "description": "Retorna la lista de usuarios",
                "operationId": "getInventory",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    }
                }
            }
        },
...
        "/user/{id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Retorna el Usuario",
                "description": "Encuentra el usuario a traves del ID",
                "operationId": "getUserById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del Usuario",
                        "required": true,
                        "type": "number",
                        "format": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "400": {
                        "description": "Id no valido"
                    },
                    "401": {
                        "description": "Token no definido"
                    },
                    "404": {
                        "description": "Elemento no encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "user"
                ],
                "summary": "Actuliza el usuario a traves del ID",
                "description": "Actualiza el Usuario",
                "operationId": "UpdateUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del Usuario",
                        "required": true,
                        "type": "number",
                        "format": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "400": {
                        "description": "Id no valido"
                    },
                    "401": {
                        "description": "Token no definido"
                    },
                    "404": {
                        "description": "Elemento no encontrado"
                    }
                }
            }
        },
...
        "/project/{id}": {
            "get": {
                "tags": [
                    "project"
                ],
                "summary": "Retorna un Proyecto",
                "description": "Retorna el proyecto",
                "operationId": "getProjectById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del proyecto",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "404": {
                        "description": "Elemento no encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "project"
                ],
                "summary": "Actuliza un proyecto a traves del ID",
                "description": "Actualiza un proyecto",
                "operationId": "UpdateProject",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del proyecto",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Category"
                            }
                        }
                    },
                    "404": {
                        "description": "Elemento no encontrado"
                    }
                }
            }
        },
        "/project": {
            "post": {
                "tags": [
                    "project"
                ],
                "summary": "Registra un proyecto",
                "description": "Registra un nuevo proyecto",
                "operationId": "CreateProject",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Elementos requeridos",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Project"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Elementos requeridos"
                    }
                }
            },
            "get": {
                "tags": [
                    "project"
                ],
                "summary": "Retorna la lista de Proyectos",
                "description": "Lista de Proyectos",
                "operationId": "getProjects",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Project"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "User": {
            "type": "object",
            "required": [
                "name",
                "lastName",
                "contactName",
                "phone"
            ],
            "properties": {
                "id": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "contactName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "User"
            }
        },
...
        "Project": {
            "type": "object",
            "required": [
                "id",
                "title",
                "idUser",
                "shortTitle",
                "idCategory",
                "description"
            ],
            "properties": {
                "id": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "shortTitle": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "idUser": {
                    "type": "number"
                },
                "idCategory": {
                    "type": "number"
                }
            },
            "xml": {
                "name": "Project"
            }
        },
        "Category": {
            "type": "object",
            "required": [
                "name",
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "Category"
            }
        },
...
        "CreateUser": {
            "type": "object",
            "required": [
                "name",
                "lastName",
                "contactName",
                "phone",
                "country",
                "city",
                "postalCode",
                "email",
                "password",
                "language"
            ],
            "properties": {
                "id": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "contactName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "active": {
                    "type": "boolean"
                },
                "country": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "postalCode": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "User"
            }
        }
    }
}